A few small changes to add compatability with hip-clang
